% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb2wave.R
\name{get_structure_nopdb}
\alias{get_structure_nopdb}
\title{Read sequence and seconday structure from csv file}
\usage{
get_structure_nopdb(filename)
}
\arguments{
\item{filename}{the csv filename, two columns with no header for three-letter aa code and ss}
}
\value{
a list with sequence, secondary structure
}
\description{
Read sequence and seconday structure from csv file
}
