% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb2wave.R
\name{get_structure_pdb}
\alias{get_structure_pdb}
\title{Read pdb extract sequence, call dssp for secondary structure, 
calculate distances to two points for sound source localisation}
\usage{
get_structure_pdb(filename, refL_resno = NULL, refR_resno = NULL)
}
\arguments{
\item{filename}{the pdb filename, only calphas are used}

\item{refL_resno}{residue number for left reference point}

\item{refR_resno}{residue number for rigth reference point}
}
\value{
a list with sequence, secondary structure and distances to R/L refs
}
\description{
Read pdb extract sequence, call dssp for secondary structure, 
calculate distances to two points for sound source localisation
}
