% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb2wave.R
\name{over}
\alias{over}
\title{Add two waves, offsetting the starting point of the second from the end of the first}
\usage{
over(a, b, offset = 0)
}
\arguments{
\item{a}{The first wave (vector of numbers)}

\item{b}{The second wave (vector of numbers)}

\item{offset}{how many indices to move the second wave with respect to the end of the first. 
Negative offset results in overlap}
}
\value{
The vector of numbers of the joint waves
}
\description{
Add two waves, offsetting the starting point of the second from the end of the first
}
\examples{
aas = get_aa_waves()
AH_wave = over(aas("A"),aas("H"))
 
}
