% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssppdb.R
\name{pdb.dssp}
\alias{pdb.dssp}
\title{DSSP call function for bio3d for old dssp version (just the --version to -V)
DSSP exec downloaded from https://swift.cmbi.umcn.nl/gv/dssp/HTML/distrib.html}
\usage{
pdb.dssp(pdb, exefile = "dssp", resno = TRUE, full = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{pdb}{pdb object read from read.pdb}
}
\description{
DSSP call function for bio3d for old dssp version (just the --version to -V)
DSSP exec downloaded from https://swift.cmbi.umcn.nl/gv/dssp/HTML/distrib.html
}
