% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb2wave.R
\name{taper}
\alias{taper}
\title{Envelopes a wave with a simple envelope.}
\usage{
taper(a, sus, decay, predecay = 0, w = 1, pad = 0)
}
\arguments{
\item{a}{The wave (vector of numbers)}

\item{sus}{How many sampling points to sustain (envelope = 1) from the first}

\item{decay}{sampling points scale for exponential decay after sustain}

\item{predecay}{second exponential summed (results in envelope > 1), for attack}

\item{w}{overall weight}

\item{pad}{shift the whole wave ahead of this many points (attempt to delay for location sensitivity)}
}
\value{
The vector of numbers of the enveloped wave
}
\description{
Envelopes a wave with a simple envelope.
}
\examples{
aas = get_aa_waves()
A_wave = taper(aas("A"),sus=10000, decay=100)
}
